requirements.txt
pandas
SQLAlchemy
pymysql
python-dotenv
seaborn
matplotlib